slack-code/requirements.txt
google-cloud-secret-manager==2.7.1
requests==2.25.1
functions-framework==2.1.0